+++ test/t/907_reg_node_two.pl
@@ -1,0 +1,82 @@
+# This is a complex test case; after creating a two node cluster on the localhost, 
+# the test case executes the commands in the Getting Started Guide at the pgEdge website.
+#
+# In this test case, we register node 2 and create the repset on that node.
+
+use strict;
+use warnings;
+use File::Which;
+use IPC::Cmd qw(run);
+use Try::Tiny;
+use JSON;
+use lib './t/lib';
+use contains;
+
+# Our parameters are:
+
+my $cmd99 = qq(whoami);
+print("cmd99 = $cmd99\n");
+my ($success99, $error_message99, $full_buf99, $stdout_buf99, $stderr_buf99)= IPC::Cmd::run(command => $cmd99, verbose => 0);
+print("stdout_buf99 = @$stdout_buf99\n");
+
+my $repuser = "@$stdout_buf99[0]";
+my $username = "lcusr";
+my $password = "password";
+my $database = "lcdb";
+my $version = "pg16";
+my $spock = "3.1";
+my $cluster = "demo";
+my $repset = "demo-repset";
+my $n1 = "~/work/nodectl/pgedge/cluster/demo/n1";
+my $n2 = "~/work/nodectl/pgedge/cluster/demo/n2";
+
+# We can retrieve the home directory from nodectl in json form... 
+my $json = `$n2/pgedge/nc --json info`;
+print("my json = $json");
+my $out = decode_json($json);
+my $homedir = $out->[0]->{"home"};
+print("The home directory is {$homedir}\n");
+
+# We can retrieve the port number from nodectl in json form...
+my $json2 = `$n2/pgedge/nc --json info pg16`;
+print("my json = $json2");
+my $out2 = decode_json($json2);
+my $port = $out2->[0]->{"port"};
+print("The port number is {$port}\n");
+
+# Register node 2 and add the replication set entry on n2: 
+
+my $cmd4 = qq($homedir/nodectl spock node-create n2 'host=127.0.0.1 port=$port user=$repuser dbname=$database' $database);
+print("cmd4 = $cmd4\n");
+my ($success4, $error_message4, $full_buf4, $stdout_buf4, $stderr_buf4)= IPC::Cmd::run(command => $cmd4, verbose => 0);
+
+print("stderr_buf4 = @$stderr_buf4\n");
+print("We just invoked the ./nc spock node-create n2 command\n");
+
+my $cmd5 = qq($homedir/nodectl spock repset-create $repset $database);
+print("cmd5 = $cmd5\n");
+my ($success5, $error_message5, $full_buf5, $stdout_buf5, $stderr_buf5)= IPC::Cmd::run(command => $cmd5, verbose => 0);
+print("We just executed the command that creates the replication set (demo-repset)\n");
+
+# Then, use the info to connect to psql and test for the existence of the replication set.
+
+my $cmd6 = qq($homedir/$version/bin/psql -t -h 127.0.0.1 -p $port -d $database -c "SELECT * FROM spock.replication_set");
+print("cmd6 = $cmd6\n");
+my($success6, $error_message6, $full_buf6, $stdout_buf6, $stderr_buf6)= IPC::Cmd::run(command => $cmd6, verbose => 0);
+print("stdout_buf6 = @$stdout_buf6\n");
+
+# Test to confirm that cluster is set up.
+
+print("We just installed pgedge/spock in $n1.\n");
+
+if(contains(@$stdout_buf6[0], "demo-repset"))
+
+{
+    exit(0);
+}
+else
+{
+    exit(1);
+}
+
+