+++ test/t/910_sub-create_n1.pl
@@ -1,0 +1,73 @@
+# This is a complex test case; after creating a two node cluster on the localhost, 
+# the test case executes the commands in the Getting Started Guide at the pgEdge website.
+#
+# In this step, we create the subscription on node 1.
+
+use strict;
+use warnings;
+use File::Which;
+use IPC::Cmd qw(run);
+use Try::Tiny;
+use JSON;
+use lib './t/lib';
+use contains;
+
+# Our parameters are:
+
+my $cmd99 = qq(whoami);
+print("cmd99 = $cmd99\n");
+my ($success99, $error_message99, $full_buf99, $stdout_buf99, $stderr_buf99)= IPC::Cmd::run(command => $cmd99, verbose => 0);
+print("stdout_buf99 = @$stdout_buf99\n");
+
+my $repuser = "@$stdout_buf99[0]";
+my $username = "lcusr";
+my $password = "password";
+my $database = "lcdb";
+my $version = "pg16";
+my $spock = "3.1";
+my $cluster = "demo";
+my $repset = "demo-repset";
+my $n1 = "~/work/nodectl/pgedge/cluster/demo/n1";
+my $n2 = "~/work/nodectl/pgedge/cluster/demo/n2";
+
+# We can retrieve the home directory from nodectl in json form... 
+my $json = `$n1/pgedge/nc --json info`;
+print("my json = $json");
+my $out = decode_json($json);
+my $homedir = $out->[0]->{"home"};
+print("The home directory is {$homedir}\n");
+
+# We can retrieve the port number from nodectl in json form...
+my $json2 = `$n1/pgedge/nc --json info pg16`;
+print("my json = $json2");
+my $out2 = decode_json($json2);
+my $port = $out2->[0]->{"port"};
+print("The port number is {$port}\n");
+
+# Then, create the subscription on node 1:
+
+my $cmd11 = qq($homedir/nodectl spock sub-create sub_n1n2 'host=127.0.0.1 port=$port user=$repuser dbname=$database' $database);
+print("cmd11 = $cmd11\n");
+my($success11, $error_message11, $full_buf11, $stdout_buf11, $stderr_buf11)= IPC::Cmd::run(command => $cmd11, verbose => 0);
+print("stdout_buf11 = @$stdout_buf11\n");
+
+# Then, we connect with psql and confirm that the subscription exists.
+
+my $cmd7 = qq($homedir/$version/bin/psql -t -h 127.0.0.1 -p $port -d $database -c "SELECT * FROM spock.subscription");
+print("cmd7 = $cmd7\n");
+my($success7, $error_message7, $full_buf7, $stdout_buf7, $stderr_buf7)= IPC::Cmd::run(command => $cmd7, verbose => 0);
+
+# Then, confirm that the subscription exists.
+
+print("We just created the subscription on ($n1) and are now verifying it exists.\n");
+
+if(contains(@$stdout_buf7[0], "sub_n1n2"))
+
+{
+    exit(0);
+}
+else
+{
+    exit(1);
+}
+