+++ test/t/912_sub-create_n2.pl
@@ -1,0 +1,72 @@
+# In this case, we create the subscription on node 2 with nc spock sub-create.
+#
+
+use strict;
+use warnings;
+use File::Which;
+use IPC::Cmd qw(run);
+use Try::Tiny;
+use JSON;
+use lib './t/lib';
+use contains;
+
+# Our parameters are:
+
+my $cmd99 = qq(whoami);
+print("cmd99 = $cmd99\n");
+my ($success99, $error_message99, $full_buf99, $stdout_buf99, $stderr_buf99)= IPC::Cmd::run(command => $cmd99, verbose => 0);
+print("stdout_buf99 = @$stdout_buf99\n");
+
+my $repuser = "@$stdout_buf99[0]";
+my $username = "lcusr";
+my $password = "password";
+my $database = "lcdb";
+my $version = "pg16";
+my $spock = "3.1";
+my $cluster = "demo";
+my $repset = "demo-repset";
+my $n1 = "~/work/nodectl/pgedge/cluster/demo/n1";
+my $n2 = "~/work/nodectl/pgedge/cluster/demo/n2";
+
+# We can retrieve the home directory from nodectl in json form... 
+my $json = `$n2/pgedge/nc --json info`;
+print("my json = $json");
+my $out = decode_json($json);
+my $homedir = $out->[0]->{"home"};
+print("The home directory is {$homedir}\n");
+
+# We can retrieve the port number from nodectl in json form...
+my $json2 = `$n2/pgedge/nc --json info pg16`;
+print("my json = $json2");
+my $out2 = decode_json($json2);
+my $port = $out2->[0]->{"port"};
+print("The port number is {$port}\n");
+
+# Create the subscription on node 2#
+
+my $cmd12 = qq($homedir/nodectl spock sub-create sub_n2n1 'host=127.0.0.1 port=$port user=$repuser dbname=$database' $database);
+print("cmd12 = $cmd12\n");
+my($success12, $error_message12, $full_buf12, $stdout_buf12, $stderr_buf12)= IPC::Cmd::run(command => $cmd12, verbose => 0);
+print("stdout_buf12 = @$stdout_buf12\n");
+
+# Then, we connect with psql and confirm that the subscription exists.
+
+my $cmd7 = qq($homedir/$version/bin/psql -t -h 127.0.0.1 -p $port -d $database -c "SELECT * FROM spock.subscription");
+print("cmd7 = $cmd7\n");
+my($success7, $error_message7, $full_buf7, $stdout_buf7, $stderr_buf7)= IPC::Cmd::run(command => $cmd7, verbose => 0);
+
+# Then, confirm that the subscription has been created.
+
+print("We just created the subscription on ($n1) and are now verifying it exists.\n");
+
+if(contains(@$stdout_buf7[0], "sub_n2n1"))
+
+{
+    exit(0);
+}
+else
+{
+    exit(1);
+}
+
+